feat: agregar soporte para instalación como paquete y mejorar la inyección de rutas en el entorno de Python
