+++ redhat_bdd_framework/__init__.py
@@ -1,0 +1,5 @@
+"""RedHat BDD Framework package."""
+
+__all__ = ["main"]
+
+from .cli import main
