+++ redhat_bdd_framework/__main__.py
@@ -1,0 +1,6 @@
+"""Module entrypoint for `python -m redhat_bdd_framework`."""
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
